unsigned long long -> uint64_t

+++ include/qb/qbtimer.h
@@ -33,12 +33,12 @@ typedef void *qb_timer_handle;
 int qb_timer_init(void (*serialize_lock) (void),
 		  void (*serialize_unlock) (void), int sched_priority);
 
-int qb_timer_add_duration(unsigned long long nanosec_duration,
+int qb_timer_add_duration(uint64_t nanosec_duration,
 			  void *data,
 			  void (*timer_fn) (void *data),
 			  qb_timer_handle * handle);
 
-int qb_timer_add_absolute(unsigned long long nanoseconds_from_epoch,
+int qb_timer_add_absolute(uint64_t nanoseconds_from_epoch,
 			  void *data,
 			  void (*timer_fn) (void *data),
 			  qb_timer_handle * handle);
@@ -51,9 +51,9 @@ void qb_timer_lock(void);
 
 void qb_timer_unlock(void);
 
-unsigned long long qb_timer_time_get(void);
+uint64_t qb_timer_time_get(void);
 
-unsigned long long qb_timer_expire_time_get(qb_timer_handle handle);
+uint64_t qb_timer_expire_time_get(qb_timer_handle handle);
 
 /* *INDENT-OFF* */
 #ifdef __cplusplus
