+++ lib/ipc_int.h
@@ -69,7 +69,7 @@ enum res_init_types {
 
 typedef struct {
 	int service __attribute__ ((aligned(8)));
-	unsigned long long semkey __attribute__ ((aligned(8)));
+	uint64_t semkey __attribute__ ((aligned(8)));
 	char control_file[64] __attribute__ ((aligned(8)));
 	char request_file[64] __attribute__ ((aligned(8)));
 	char response_file[64] __attribute__ ((aligned(8)));
